+++ include/boost/sort/pdqsort/pdqsort.hpp
@@ -127,7 +127,7 @@ namespace pdqsort_detail {
                 while (sift != begin && comp(tmp, *--sift_1));
 
                 *sift = BOOST_PDQSORT_PREFER_MOVE(tmp);
-                limit += cur - sift;
+                limit += static_cast<int>(cur - sift);
             }
         }
 
@@ -262,7 +262,7 @@ namespace pdqsort_detail {
         }
 
         int l_size = 0, r_size = 0;
-        int unknown_left = (last - first) - ((num_r || num_l) ? block_size : 0);
+        int unknown_left = static_cast<int>((last - first) - ((num_r || num_l) ? block_size : 0));
         if (num_r) {
             // Handle leftover block by assigning the unknown elements to the other block.
             l_size = unknown_left;
